# documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Homework-9",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "http://localhost:4000/register",
        "operationId": "http://localhost:4000/register",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/httplocalhost4000registerrequest"
                  },
                  {
                    "example": {
                      "id": 103,
                      "email": "daniadjtnk@mail.com",
                      "gender": "Female",
                      "password": "12345qwer",
                      "role": "Tech Educator"
                    }
                  }
                ]
              },
              "example": {
                "id": 103,
                "email": "daniadjtnk@mail.com",
                "gender": "Female",
                "password": "12345qwer",
                "role": "Tech Educator"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "http://localhost:4000/register1",
        "operationId": "http://localhost:4000/register1",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/httplocalhost4000registerrequest1"
                  },
                  {
                    "example": {
                      "email": "daniadjtnk@mail.com",
                      "password": "12345qwer"
                    }
                  }
                ]
              },
              "example": {
                "email": "daniadjtnk@mail.com",
                "password": "12345qwer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
...
      "httplocalhost4000registerrequest": {
        "title": "httplocalhost4000registerrequest",
        "required": [
          "id",
          "email",
          "gender",
          "password",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "example": {
          "id": 103,
          "email": "daniadjtnk@mail.com",
          "gender": "Female",
          "password": "12345qwer",
          "role": "Tech Educator"
        }
      },
      "httplocalhost4000registerrequest1": {
        "title": "httplocalhost4000registerrequest1",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "daniadjtnk@mail.com",
          "password": "12345qwer"
        }
      }
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}